cleanup old pre java 8 code

+++ src/main/java/dan200/computercraft/shared/util/RedstoneUtil.java
@@ -31,7 +31,7 @@ public class RedstoneUtil {
             IBlockState state = world.getBlockState(pos);
             if (block == Blocks.REDSTONE_WIRE) {
                 if (side != EnumFacing.UP) {
-                    power = ((Integer) state.getValue(BlockRedstoneWire.POWER)).intValue();
+                    power = state.getValue(BlockRedstoneWire.POWER).intValue();
                 } else {
                     power = 0;
                 }
